Remove required mod_autoindex
The module is NOT required for TYPO3. In fact, in the
default TYPO3 .htaccess directory listings are disabled:

`<IfModule mod_autoindex.c>`
        Options -Indexes
`</IfModule>`

The module mod_autoindex is not required to disable the
directoriy listings. It provides the directory listings
in the first place (which can then be disabled).

So, if one wants to add the module, they can, but it is
not a requirement for TYPO3.