Clarify section on Apache modules
The previous text was slightly misleading because not all modules
are a requirement for TYPO3. But the current modules are used
in the default .htaccess. Addressing this and adding a link
to the .htaccess file in the core makes it more clear and makes
it easier for people to followup and check the requirements
and how they are used - even if they have not yet installed
TYPO3.