Use proven DDEV install technique that uses composer --no-install
The existing install doesn't work if mutagen is enabled, I'm not sure why. 

But the DDEV technique on https://ddev.readthedocs.io/en/latest/users/quickstart/#typo3-composer-build
always works.

Note that this also changes the `typo3/cms-base-distribution:^11` to just `typo3/cms-base-distribution` so it doesn't have to be maintained on next release.