Remove "Tuning TYPO3" page
The information about Opcache is reduced to only the settings
which are TYPO3 specific. This significantly reduces the text
and now adds it to the System Requirements page (where people
are more bound to look for basic system requirements).

The page was initially created when the information about opcache
was added. Now, it only includes information about Opcache and most
of that just uses the default Opcache settings. So the page does
not provide that much value and the topic "Performance tuning"
is slightly out of place in a "Getting started" tutorial, it is
more an advanced topic and should be handled more in depth (if
handled at all on docs.typo3.org).

It might be nice to add a page for performance tuning and / or
information for setting up TYPO3 for system administrators on a
more advanced level, but the question is if this is in the right
place in the "Getting Started" guide.

Resolves: #233