[!!!][TASK] Delete russian main version
There have been no changes applied for the version 13 related changes. And it is not expected that the current changes will be applied to the main Russian version either.

We will keep the 12.4 version but there is no active development at the moment.

We discussed this in the team meeting, 10.10.2024

Releases: main