[BUGFIX] Mention that site config is auto generated for root pages (#715)
Updates the "Site Management" page with an attention admonition informing about the automated generation of page configurations when you create a page in the root of the page tree. Also adds another section header about editing the site configuration in order to keep creating and editing the site configuration as separate steps.